# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "XML DB Import Project Documentation",
        "description": "API Documentation for XML db import",
        "contact": {
            "email": "delvinnj02@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/register": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Create a new user with bearerAuth",
                "operationId": "registerNewUser",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "name",
                                        "type": "string"
                                    },
                                    "email": {
                                        "description": "email",
                                        "type": "string"
                                    },
                                    "password": {
                                        "description": "password",
                                        "type": "string"
                                    },
                                    "password_confirmation": {
                                        "description": "Confirm password",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error occurred",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        }
    },
...
    "components": {
        "securitySchemes": {
            "Bearer": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer *********************)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}